Upload compounds to Workbench from Target Search
This commit contains a limited implementation of (see title). Currently
it's only applied to the Annotation table. I'll likely templatize the
Annotation and Activity tables in the near future, since new_ts.html is
getting insanely long at this point, and there's quite a lot of stuff
that can be deduplicated.