Add IDs to search boxes via row selection feature.
Users can now select rows and send the compound/target IDs to the
corresponding search box with a button similar to the workbench compound
submission process.

Also I disabled duplicate CID handling. IDs need to be exact to be
useful in most contexts. Later on I'll put in a feature to merge
duplicate records.