Rewrote drug-indication table code
The previous code stored drug-indication tables as bare HTML code in a
JS variable, which can get quite inefficient. Further, every table for
each compound was rendered on page load.

Now, the table data is stored in JSON format, which saves significant
space, and might yield some flexibility benefits down the road. Also,
the DataTable is now rendered on-demand and destroyed when the modal
window is dismissed.