Added a precheck routine for Target Search compound import.
The precheck routine checks for compounds that do not have structure
information in the ChEMBL database, as well as for compounds that are
already in the user's workbench. If we discover other conditions that
could cause an import to fail, we can easily add further restrictions.

Also, chembl_helpers.py has been renamed to simply helpers.py to better
reflect its current purpose, and for consistency with the new helpers.py
in myCompounds.

In targetsearch/helpers.py, get_chembl_sdfs() and get_chembl_smiles()
will now cast chemblIds into a tuple, so we can now pass in lists, sets,
etc. as is. Hopefully this doesn't create problems down the road...